+++ R/plotMethods.R
@@ -4,9 +4,8 @@
 #'
 #' @description \code{plot.gsmarpred} is plot method for class 'gsmarpred' objects
 #'
+#' @inheritParams predict.gsmar
 #' @param x object of class \code{'gsmarpred'} created with \code{predict.gsmar}.
-#' @param nt a positive integer specifying the number of observations to be plotted
-#'   along with the prediction. Default is \code{round(length(data)*0.2)}.
 #' @param add_grid should grid a be added to the plots?
 #' @param ... arguments passed to function \code{grid}.
 #' @details This method is intended for plotting forecasts of GSMAR processes.
@@ -22,12 +21,24 @@
 #' @export
 
 plot.gsmarpred <- function(x, ..., nt, add_grid=TRUE) {
+
   gsmarpred <- x
   data <- as.ts(gsmarpred$gsmar$data)
   n_obs <- length(data)
   q <- gsmarpred$q
+  mix_weights <- gsmarpred$mix_weights
+  M <- sum(gsmarpred$gsmar$model$M)
+  n_mix <- nrow(gsmarpred$gsmar$mixing_weights)
+
+  # Graphical parameters
+  old_par <- par(no.readonly=TRUE) # Save old settings
+  on.exit(par(old_par)) # Restore the settings before quitting
+  if(mix_weights) {
+    par(mfrow=c(2, 1), mar=c(2.6, 2.6, 2.6, 2.6))
+  }
+
   if(missing(nt)) {
-    nt <- round(length(data)*0.2)
+    nt <- round(length(data)*0.15)
   } else {
     stopifnot(nt > 0 & nt %% 1 == 0)
     if(nt > length(data)) {
@@ -35,21 +46,48 @@ plot.gsmarpred <- function(x, ..., nt, add_grid=TRUE) {
       nt <- length(data)
     }
   }
-  make_ts <- function(a) ts(c(data[n_obs], a), start=time(data)[n_obs], frequency=frequency(data))
+  make_ts <- function(a, mix=FALSE, m) { # a is vector of values; if mix == TRUE, give component m in 1,...,M
+    last_obs <- ifelse(mix, gsmarpred$gsmar$mixing_weights[n_mix, m], data[n_obs])
+    ts(c(last_obs, a), start=time(data)[n_obs], frequency=frequency(data))
+  }
   ts_pred <- make_ts(gsmarpred$pred)
   ts_dat <- ts(data[(n_obs - nt + 1):n_obs], start=time(data)[n_obs - nt + 1], frequency=frequency(data))
-  t0 = time(ts_pred)
+  t0 <- time(ts_pred)
   all_val <- c(ts_dat, ts_pred, gsmarpred$pred_ints)
 
   if(gsmarpred$pi_type == "two-sided") {
-    ts1 <- lapply(1:(length(q)/2), function(i1) make_ts(gsmarpred$pred_ints[,i1]))
-    ts2 <- lapply((length(q)/2 + 1):length(q), function(i1) make_ts(gsmarpred$pred_ints[,i1]))
+    ts1_lapply <- function(pred_ints, mix=FALSE, m) lapply(1:(length(q)/2), function(i1) make_ts(pred_ints[,i1], mix, m)) # Lower bounds
+    ts2_lapply <- function(pred_ints, mix=FALSE, m) lapply((length(q)/2 + 1):length(q), function(i1) make_ts(pred_ints[,i1], mix, m)) # Upper bounds
+
+    ints1 <- gsmarpred$pred_ints
+    ints1_mix <- gsmarpred$mix_pred_ints
 
   } else {
     what_to_rep <- ifelse(gsmarpred$pi_type == "upper", round(min(all_val)) - 3, round(max(all_val)) + 3) # Otherwise "lower" or "none"
-    ts0 <- rep(what_to_rep, times=length(ts_pred))
-    ts1 <- lapply(seq_along(q), function(i1) make_ts(ts0)[-1]) # Upper or lower graphical device box bound
-    ts2 <- lapply(seq_along(q), function(i1) make_ts(gsmarpred$pred_ints[,i1])) # Upper or lower prediction bound
+    what_to_rep_mix <- ifelse(gsmarpred$pi_type == "upper", 0, 1)
+
+    ts1_lapply <- function(pred_ints, mix=FALSE, m) lapply(seq_along(q), function(i1) make_ts(pred_ints[,i1], mix, m)[-1]) # Upper or lower graphical device box bound (redundant argument)
+    ts2_lapply <- function(pred_ints, mix=FALSE, m) lapply(seq_along(q), function(i1) make_ts(pred_ints[,i1], mix, m)) # Make upper or lower prediction bound
+
+  #  ts0 <- rep(what_to_rep, times=length(ts_pred))
+  #  ts1 <- lapply(seq_along(q), function(i1) make_ts(ts0)[-1]) # Upper or lower graphical device box bound
+  #  ts2 <- lapply(seq_along(q), function(i1) make_ts(gsmarpred$pred_ints[,i1])) # Upper or lower prediction bound
+
+    ints1 <- matrix(rep(what_to_rep, times=length(q)*length(ts_pred)), ncol=length(q))
+    ints1_mix <- array(rep(what_to_rep_mix, times=length(ts_pred)*length(q)*M), dim=c(length(ts_pred), length(q), M))
+  }
+
+  # Create the ts objects for confidence intervals
+  ts1 <- ts1_lapply(ints1)
+  ts2 <- ts2_lapply(gsmarpred$pred_ints)
+
+  if(mix_weights) {
+    ts1_mw <- vector(mode="list", length=M) # Sublist for each regime
+    ts2_mw <- vector(mode="list", length=M)
+    for(m in 1:M) {
+      ts1_mw[[m]] <- ts1_lapply(ints1_mix[, , m], mix=TRUE, m=m)
+      ts2_mw[[m]] <- ts2_lapply(gsmarpred$mix_pred_ints[, , m], mix=TRUE, m=m)
+    }
   }
 
   ts.plot(ts_dat, ts_pred, gpars=list(col=c("black", "blue"), lty=1:2,
@@ -58,12 +96,42 @@ plot.gsmarpred <- function(x, ..., nt, add_grid=TRUE) {
                                       main=paste("Forecast", gsmarpred$n_ahead, "steps ahead")))
   if(add_grid) grid(...)
 
+  # Draw the prediction intervals
+  draw_poly <- function(ts1_or_ts2, pred_ts, col) polygon(x=c(t0, rev(t0)), y=c(ts1_or_ts2, rev(pred_ts)), col=col, border=NA)
+  col_pred <- grDevices::rgb(0, 0, 1, 0.2)
   if(gsmarpred$pi_type %in% c("two-sided", "upper", "lower")) {
     for(i1 in 1:length(gsmarpred$pi)) { # Go through the prediction intervals
-      polygon(x=c(t0, rev(t0)), y=c(ts1[[i1]], rev(ts_pred)), col=grDevices::rgb(0, 0, 1, 0.2), border=NA)
-      polygon(x=c(t0, rev(t0)), y=c(ts2[[i1]], rev(ts_pred)), col=grDevices::rgb(0, 0, 1, 0.2), border=NA)
+      draw_poly(ts1[[i1]], ts_pred, col=col_pred)
+      draw_poly(ts2[[i1]], ts_pred, col=col_pred)
     }
   }
+
+  # Plot mixing weight forecasts
+  if(mix_weights) {
+
+    # Point forecasts
+    colpal_mw <- grDevices::colorRampPalette(c("blue", "turquoise1", "green", "red"))(M)
+    colpal_mw2 <- grDevices::adjustcolor(colpal_mw, alpha.f=0.5)
+    mix_ts <- ts(gsmarpred$gsmar$mixing_weights[(n_mix - nt + 1):n_mix,], start=time(data)[n_obs - nt + 1],
+                 frequency=frequency(data))
+    mix_pred_ts <- ts(rbind(mix_ts[nrow(mix_ts),], gsmarpred$mix_pred), start=time(data)[n_obs], frequency=frequency(data))
+    ts.plot(mix_ts, mix_pred_ts, gpars=list(col=c(colpal_mw2, colpal_mw), ylim=c(0, 1), lty=c(rep(1, M), rep(2, M))))
+    legend("topleft", legend=paste0("regime ", 1:M), bty="n", col=colpal_mw, lty=1, lwd=2,
+           text.font=2, cex=0.65, x.intersp=0.5, y.intersp=1)
+    if(add_grid) grid(...)
+
+    # Individual prediction intervals as for the mixing weights
+    colpal_mw3 <- grDevices::adjustcolor(colpal_mw, alpha.f=0.2)
+    if(gsmarpred$pi_type %in% c("two-sided", "upper", "lower")) {
+      for(m in 1:M) { # Go through regimes
+        for(i1 in 1:length(gsmarpred$pi)) { # Go through the prediction intervals
+          draw_poly(ts1_mw[[m]][[i1]], mix_pred_ts[, m], col=colpal_mw3[m])
+          draw_poly(ts2_mw[[m]][[i1]], mix_pred_ts[, m], col=colpal_mw3[m])
+        }
+      }
+    }
+  }
+
   invisible(gsmarpred)
 }